test(kerberos): remove crufty tests, replace with working ones
