+++ package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "chai": "^4.1.2",
+    "chai-string": "^1.4.0",
     "clang-format": "^1.2.3",
     "eslint": "^4.19.1",
     "eslint-plugin-prettier": "^2.6.0",
